+++ src/pages/mine/components/AboutList/index.module.less
@@ -114,7 +114,7 @@
   }
   .intro__img {
     display: flex;
-    height: 150px;
+    height: 175px;
     background: #fff;
     margin: 30px 20px 10px;
     border-radius: 10px;
