+++ src/pages/mine/mine.config.ts
@@ -1,6 +1,6 @@
 export default definePageConfig({
   navigationBarTitleText: '我的',
   enablePullDownRefresh: false,
-  enableShareAppMessage: true,
-  enableShareTimeline: true
+  // enableShareAppMessage: true,
+  // enableShareTimeline: true
 })
